+++ massiv/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.5.3
 
 * Fix `tanA` and `tanhA`. [#96](https://github.com/lehins/massiv/pull/96)
+* Add `scons` and `sunconsM`
+* Relax argument of `snoc` and `cons` constraint to `Load` vectors
 
 # 0.5.2